Add kernel requires to kernel-signed requires (#90)
* kernel-signed-x64: Add missing requires

The %post step fails because /sbin/depmod is not present. depmod
is supplied by the kmod package.

This error manifested as a hyper-v boot hang where the image is
stuck infinitely waiting for the rootfs to mount. Since depmod was
never run during kernel installation, the module database is stale. Then when the initramfs regeneration occurs, certain modules (i.e. hv_storvsc)
are not available for dracut to include into the initrd.

Bump release number

* kernel-signed-aarch64: Add missing requires

The %post step fails because /sbin/depmod is not present. depmod
is supplied by the kmod package.

This error manifested as a hyper-v boot hang where the image is
stuck infinitely waiting for the rootfs to mount. Since depmod was
never run during kernel installation, the module database is stale. Then when the initramfs regeneration occurs, certain modules (i.e. hv_storvsc)
are not available for dracut to include into the initrd.

Bump release number

* kernel: Bump release

* kernel: clean up lingering invalid aarch64 configs

* kernel: Fix bogus date rpmlint message