[main] Enable NO_IOMMU kernel flag. (#2385)
* [main] Enable NO_IOMMU kernel flag.

In order to support high data throughput for network connections, user-mode
drivers require direct access to the underlying devices. Such access can be
managed/protected by the IOMMU controller if it is present in the hardware.

For VMs, we may not have IOMMU exposed by the hypervisor. So we must provide
a way for the VFIO module to say that there is no IOMMU but still use VFIO
to keep the software stack the same.

The Mariner kernel today is compiled such that IOMMU is always required for
such user-mode drivers to work. This is a problem for virtual machines where
such requirement cannot be met.

This fix changes the kernel NO_IOMMU compile option such that it is possible
for the root user to choose whether to require IOMMU or not based on their
needs without recompiling the kernel.

The default is that IOMMU is required.

The root user must explicitly disable the IOMMU requirement with the following
command:
/sbin/modprobe vfio enable_unsafe_noiommu_mode=Y

* Enable NO_IOMMU: update version of kernel related spec files.

* Enable NO_IOMMU: update toolkit references.

* Enable NO_IOMMU: update pkggen references.

* Update kernel signature file after merge with main.