installkernel: Fix mariner.cfg symlink (#2696)
* installkernel: Fix mariner.cfg symlink (#2647)

The installkernel script is generating the /boot/mariner.cfg symlink
that required the presence of /boot, which is not guaranteed.

ln -sf "$DEST_DIR"/linux-"$KERNEL_VERSION".cfg "$DEST_DIR"/mariner.cfg

where $DEST_DIR can be /boot. When there is a dedicated /boot partition,
once grub loads, it will try to load mariner.cfg, which would be a
broken path to /boot/linux-<version>.cfg. This leads to the grub error
message:

	error: not a regular file

Instead the symlink should be relative to the current directory. So
remove the reference to $DEST_DIR in the symlink generation step.

Signed-off-by: Chris Co <chrco@microsoft.com>