Bug fix for `ResolveCompetingPackages` when building on RPM-based distro. (#3223)
The current implementation of `ResolveCompetingPackages` assumes it is running on a
distro that doesn't use RPM. Specifically, it doesn't set the root directory which
allows for confusion between the packages being built for CBL-Mariner and the packages
installed on the build system. This change fix this problem by setting the root
directory when calling the `rpm` command.