Fix ISO mount retry race condition (#5578)
* Update runliveinstaller to exit mount retry early

There is a race condition where we check if the cdrom mount is created, if not we go to mount it with `retry`. But the mount might be established during that period, and `mount` returns an error code if its target is already mounted so the retry will never exit early. Switch to using `eval` so we can pass a more complex bash command: `{ <check> || mount... ; }`.