toolkit: add feature to profile & trace to have better diagnostics (#5640)
This PR introduces 3 additional flags, by default all the three flags are disabled.

- ENABLE_TRACE: Setting this flag to y, instruments the various tools in the
toolkit to record trace data. Tracing is a way to instrument code to analyze
latency throughout the lifecycle of a call or user request. Traces provide an
overview of how much latency each component contributes to the overall latency
in a system. Traces can span multiple Go processes.
- ENABLE_MEM_PROFILE: Setting this flag to y instruments the toolkit to record
the memory profile.
- ENABLE_CPU_PROFILE: Setting this flag to y instruments the toolkit to record
the cpu profile. 

Profiling tools analyze the complexity and costs of a Go program such as its
memory usage and frequently called functions to identify the expensive sections
of a Go program.

It is advisable to only turn one of the flags for a run to get data that does
not have noise due to other profiling/tracing flags.

The trace/profile data is created in the build/profile directory.
To use the profile data: `go tool pprof -http=:8080 file.profile`
To use the trace data: `go tool trace -http=:8080 file.trace`

Reference: https://go.dev/doc/diagnostics    
Signed-off-by: Muhammad Falak R Wani <falakreyaz@gmail.com>