Sparse raw image. (#5950)
Use a sparse raw image file for the intermediate build image file. This will allow images to be built that have a larger capacity than the free disk space available on the build machine (so long as the committed size is within the free disk space). This will also speed up the build a little bit since it will no longer be pointlessly writing zeros to disk.

Changes:

- Use sparse image file in imager utility.