Image config: Fix for when no user password is specified. (#6443)
Currently, the situation where a user is specified in the new-image config or the image-customizer config without a password is not handled. This change ensures that the user password login is disabled in such situations (by either not passing a password value to useradd or by setting the password value to ! in the /etc/shadow file).

In addition, this change fixes a bug that could allow the plain-text password to be printed when the trace logs are enabled.