Tobiasb/openssl remove arch flag (#6912)
The upstream openssl.spec has quite a bit of logic to figure out the correct architecture to pass to Configure, and figure out a couple flags. However, we only support two architectures and they're built on the target system, which Configure will figure out.

This change removes the logic and hard-codes the only flag that was relevant.