feat: change default hidepid setting to be opt-in (#6969)
Change the user-facing option HidepidDisabled to EnableHidepid to allow user to now opt-in to add hidepid settings.

This change inverts the default setting when no option is supplied to configure hidepid=0, which is more generally desired for better software compat for applications and system services. User can now opt-in to configure the tools to set up hidepid=2.
 
This change also simplifies the hidepid code by only accessing the option from the config right before use. This allows us to remove unnecessary function arguments passed through the various call layers.

Signed-off-by: Chris Co <chrco@microsoft.com>