Add SymCrypt and SymCrypt-openssl to images with openssl (#7977)
Add SymCrypt and SymCrypt-openssl as Recommends packages for openssl and add them to images with openssl.

In AZL 3 openssl relies on SymCrypt for FIPS mode. As such, we want SymCrypt and the related SymCrypt-openssl packages installed where openssl is installed. To accomplish this, we do two things:

Add these packages as Recommends in the openssl spec. This means that if the package manager can find these packages, it will install them when openssl is installed, but it won't refuse to install openssl if it can't find them.
Add these packages to any image that has openssl.

One may wonder why we didn't simply add these packages as Requires to openssl. The issue here is that SymCrypt has a BuildRequires on python, which in turn requires openssl, so there's a circular dependency.