containerized-rpmbuild: fix tdnf --enablerepo value for 3.0 (#8404)
A recent rebranding change #8238 changed the repo name used by the containerized-rpmbuild container for mariner to azl-3.0-daily-build, but didn't change the default tdnf argument, which continued to use mariner-3.0-daily-build. This results in tdnf not working in the container.

This fix updates the tdnf argument, basing it on the string in the repo file, so we don't have to copy/paste it later.