Fix shadow-utils to reenable user groups (#9363)
Change #7761 upgraded shadow-utils and took part of a change from upstream that modifies /etc/login.defs so that USERGROUPS_ENAB is set to no. This makes all users be created in the same group (in our case users) rather than created with their own group.

This change undoes that, so users will once again be created with their own group with the same name as the user.