irqbalance: Tweak irqbalance environment file to squelch systemd warning (#9548)
Our 3.0 images produce some extra noise during boot:
irqbalance: "Referenced but unset environment variable evaluates to an empty string: IRQBALANCE_ARGS"

This is due to the variable IRQBALANCE_ARGS, referenced in irqbalance.service, not being defined in the EnvironmentFile pointed to by that service. Sed the Environment file to uncomment IRQBALANCE_ARGS, explicitly setting it to the empty string.