Toolkit: Do not give GPT partitions a default label of "primary". (#9932)
When creating partitions using `parted`, the API is:

- MBR: `parted mkpart <partition-type> <fs-type> ...`
- GPT: `parted mkpart <partition-label> <fs-type> ...`

But currently we are treating the first param as always being the `<partition-type>`, when for GPT it is `<partition-label>`. This has the effect that GPT partitions are given a default label of "primary", which is a tad silly. (If an explicit label is provided by the user, then the label is overridden in a subsequent step.)

This change fixes this behavior so that GPT partitions have a default label of nothing ("").