fix: upgrade walinuxagent to 2.11.1.4 and add azurelinux patch (#10016)
Upgrade WALinuxAgent to the most recent release - 2.11.1.4.

Add patch to include azurelinux support into the agent. The patch should
be removed in the next 2.12 update.

Also because WALinuxAgent will automatically update at runtime, thus
removing the patched-in support for azurelinux, add another patch to
bump the version to 2.11.8.8 so further 2.11 hotfixes are not taken.
Since 2.12 will have azurelinux support, the agent will be able to
automatically upgrade to 2.12 at runtime.

fix: correct changelog date warning

fix: use /usr/bin and /usr/lib/systemd paths

In the upstream code's setup.py, the bin path and systemd path are set
through calls into the distro's osutil class. In our case, since we pass
--lnx-distro='mariner', the code will use the MarinerOsUtil derived
class for these paths, which are /usr/bin and /usr/lib/systemd
respectively.

In addition, now that the paths are set correctly again, remove the sed
operation to the waagent.service so the waagent path points to the
correct location again.

Also fix the ephemeral-disk-warning.service to use the proper path as
well.

Signed-off-by: Chris Co <chrco@microsoft.com>