Image Customizer: Allow omitting disk maxSize and partition start. (#10383)
Allow the partition start to be inferred from the previous partition's end. Also, allow the disk's maxSize to be inferred from the size/end of the last partition.

In addition, since the partition start can now be omitted, require the partitions to be specified in order. Fortunately, most users do this anyway.