Enable usb hiddev and serial ch341 (#10368)
Enable USB HID support for AMD. This feature allows devices to use the HIDDEV API. "The two big examples for this are power devices (especially uninterruptible power supplies) and monitor control on higher end monitors."
Enable the module ch341. This is being added as an optional loadable module to support using a Winchiphead CH341 single port USB to serial adapter.