update wget to fix potential infinite loop (#10461)
With our configuration of wget2, which uses the openssl tls provider, we can get into an infinite loop with certain error conditions. The issue is that for both SSL_read and SSL_open, a return value of 0 is an error but the code treats it as success.

This is tracked upstream with this issue.