Disable xen debugfs, and I2C Baytrail configs (#10364)
Disable the following:
xen debugfs- Xen is not officially supported by Azure Linux, therefore remove the building and loading of debugfs. As noted in the Kconfig, "Enabling this option may incur a significant performance overhead" .
I2C Baytrail config - This config was added during initial CBl-Mariner creation and not by specific request. The Baytrail hardware does not appear to be a supported scenario. Therefore remove from being built in to reduce kernel size.