Image Customizer: Change additionalFiles to a list. (#10517)
Change `additionalFiles` from a map to a list. This will ensure that the copy operation has a fixed and consistent ordering, instead of being randomized by the map's hash algorithm.

In addition, add a feature that allows the content of additional files to be specified inline with in the YAML config.