make install_dependencies work with file dependencies (#11054)
In a containerized-rpmbuild environment, the function install_dependencies will install all BuildRequires dependencies from a spec file. However, it wasn't able to handle file dependencies like BuildRequires: %{_bindir}/cmp, which requires looking up the package that provides the file.

This change uses tdnf repoquery to get the appropriate package.

I also added usage of jq in that function, so I added it to the list of packages we always install (it's useful anyway), and did a minor refactor of how we do that.