Build PCI_HYPERV as a built-in (#11992)
Change PCI_HYPERV to be builtin. This also requires PCI_HYPERV_INTERFACE to be builtin as well. PCI_HYPERV was previously loaded as a module but it results in race conditions when devices are mapped and unmapped. AzureLinux is already starting to access PCIe Devices, when the MMIO mappings from UEFI haven't been fully torn down yet. That causes conflicts, which shows up as missing disks to the customer. If AzureLinux has hyperv pci built into the kernel, then that changes the load ordering of the drivers in the kernel, which has the desired side-effect that UEFI mappings get torn down before the AzureLinux creates net mappings.