Add ipmitool support to kernel-64k (#12154)
Enable support for ipmitool. ipmitool is available through the ipmitool package on AzureLinux. The config changes in this PR ensure the devices created are visible in /dev/. From the ipmitool project: "ipmitool is a utility for managing and configuring devices that support the Intelligent Platform Management Interface. IPMI is an open standard for monitoring, logging, recovery, inventory, and control of hardware that is implemented independent of the main CPU, BIOS, and OS."

Co-authored-by: Thomas Ballasi <tballasi@linux.microsoft.com>