readline: add patch to fix crash when readline is started with an invalid locale specification (#12903)
readline 8.2 would crash when _rl_init_locale is called with an invalid locale specification (e.g locale not available in the system). This causes a problem in initramfs emergency shell because it was not packaged with the preferred locale (C.UTF-8), thus bash will crash and no emergency shell is available. This issue does not affect Mariner 2 as it is using readline 8.1, and the bug seems to be introduced by version 8.2.

Backport patch to address readline 8.2 crash when started with an invalid locale specification.

Signed-off-by: Thien Trung Vuong <tvuong@microsoft.com>