Bphilip/refactor version entangle check add release number (#12910)
MOFED and associated kernel modules did not include kernel minor version in the entangle check. check_entangle script did not provide enough details when it failed.

Added a unit test to make sure the script was indeed doing everything as expected. There was some refactoring done to the script to make the workflow more compact.

Co-authored-by: Binu Jose Philip <bphilip@microsoft.com>